requirements.txt
PyYAML==3.11
argparse==1.2.1
requests==2.7.0
wsgiref==0.1.2
